feat(front): Organizando importações da tipagem
